read file from command line argument
Accept a filename as the single argument on the command line and read
(and, for the moment, print) it.

This is derived from an example given on argument handling[1], which was
released under a simple permissive license[2] that allows "any person...
to deal in the Work without restriction, including without limitation
the rights to use, copy, modify, merge, publish, distribute, sublicense,
and/or sell copies of the Work".

[1] Tutorials/Programming Haskell/Argument handling, available at
http://www.haskell.org/haskellwiki/index.php?title=Tutorials/Programming_Haskell/Argument_handling&oldid=54615

[2] HaskellWiki:Copyrights, available at
http://www.haskell.org/haskellwiki/index.php?title=HaskellWiki:Copyrights&oldid=1590

Issue #1 parse downloaded song list