render songs as comma separated values
Create a simple rendering scheme that will print a header in each field
for the field name and then render each song as a list of fields
separated by commas. Such a file is suitable for importing into most
spreadsheet software, where it can be sorted, formatted, and printed.

Close #3 print comma separated list of songs