fetch song list from the web
Instead of parsing a file on disk, hardcode the URI of the different
song lists, parse the command line arguments to figure out which one to
use, and then download it.

Also simplify and restrict imports to reduce clutter and repeated
namespaces, add documentation to a few sections to clarify the code,
clean up whitespace to make the code more readable, and convert several
function compositions of the form "(a . b) c" to "a . b $ c" to better
follow idiomatic Haskell.

Close #5 download song list